Refactored mvn_bridge
